// packages/command-manager/demo/interactors/structure/type.ts
// 同步 DisplayControll 的事件
const eventsType = ['click', 'dblclick', 'wheel',
    'touchstart', 'touchend', 'touchmove',
    'mouseup', 'mousedown', 'mousemove',
    'mouseover', 'mouseout',
    'keydown', 'keyup',
    'dragstart', 'dragmove', 'dragend'
];

export enum EventType {
    CLICK = 'click',
    MOUSEOVER = 'mouseover',
    MOUSEOUT = 'mouseout',
    MOUSEMOVE = 'mousemove',
    MOUSEUP = 'mouseup',
    MOUSEDOWN = 'mousedown',
    DRAGSTART = 'dragstart',
    DRAGMOVE = 'dragmove',
    DRAGEND = 'dragend',
}